how often property trasformed to int

+++ lib/src/disease_model/medicine_model.dart
@@ -22,7 +22,7 @@ class Medicine {
   String? activeSubstance;
 
   @JsonKey(name: 'how_often')
-  double? howOften;
+  int? howOften;
 
   @JsonKey(name: 'how_many')
   double? howMany;
